Combined `sound` package back into `sound.py`
